+++ res/themes/AlohaBlue/css/AlohaLogin.scss
@@ -1,0 +1,22 @@
+/*
+ * WeHeal.org login's styles
+ */
+
+.mx_Login {
+    display: block;
+    width: 100%;
+    height: 100%;
+    margin: 0;
+    padding: 20px;
+    background: transparent url(../../themes/weheal/img/background2.jpg) 50% 50%;
+    background-size: cover;
+    overflow: auto;
+}
+
+.mx_Login_box {
+    float: right !important;
+    min-height: 88%;
+    margin: 0 100px 0 0;
+    padding: 30px 40px;
+    background-color:rgb(255, 255, 255, 0.65);
+}
