+++ res/themes/AlohaBlue/css/_AppsDrawer.scss
@@ -1,0 +1,4 @@
+
+.mx_AppsDrawer {
+    margin: 0;
+}
