+++ res/themes/AlohaBlue/css/_BaseAvatar.scss
@@ -1,0 +1,4 @@
+.mx_BaseAvatar_initial {
+    background-color:$accent-color;
+    border-radius: 50%;
+}
