+++ res/themes/AlohaBlue/css/_EventTile.scss
@@ -1,0 +1,4 @@
+.mx_EventTile_line {
+    margin-top: 3px;
+    margin-left: 8px;
+}
