+++ res/themes/AlohaBlue/css/_FIlePanel.scss
@@ -1,0 +1,4 @@
+.mx_FilePanel .mx_EventTile_line {
+    margin-top: 4px;
+    margin-left: 6px;
+}
