+++ res/themes/AlohaBlue/css/_LeftPanel.scss
@@ -1,0 +1,8 @@
+
+.mx_LeftPanel .mx_BottomLeftMenu {
+    border-top: 1px solid rgb(186,221,246, 0.9);
+}
+
+.mx_TagPanel .mx_TagPanel_divider { 
+    border-bottom: 1px solid rgb(186,221,246, 0.8);
+}
