+++ res/themes/AlohaBlue/css/_MatrixToolbar.scss
@@ -1,0 +1,11 @@
+.mx_MatrixToolbar {
+  background-color: $accent-color;
+  color: $accent-fg-color;
+}
+
+.mx_MatrixToolbar_content {
+  font-size: 12px;
+  span {
+    font-size: 12px;
+  }
+}
