+++ res/themes/AlohaBlue/css/_RoomSubList.scss
@@ -1,0 +1,15 @@
+
+.mx_RoomSubList_labelContainer {
+    padding-top: 10px;
+}
+
+.mx_RoomSubList_label {
+    height: 24px;
+    padding-top: 4px;
+    padding-bottom: 8px;
+    font-size: 14px;
+}
+
+.mx_RoomSubList_roomCount {
+    font-size: 14px;
+}
