+++ res/themes/AlohaBlue/css/_SearchBox.scss
@@ -1,0 +1,4 @@
+
+.mx_SearchBox {
+    margin-bottom: 20px;
+}
