+++ res/themes/AlohaBlue/css/_common.scss
@@ -1,0 +1,19 @@
+body {
+    color:#111;
+}
+
+.mx_HomePage {
+ max-width: 1140px;
+}
+
+.mx_HomePage_toprow {
+    padding: 20px 0;
+}
+
+.mx_HomePage_row {
+    padding-bottom: 30px;
+    border-bottom: 1px solid rgb(186,221,246, 0.6);
+}
+
+
+
