+++ res/themes/AlohaBlue/css/main.scss
@@ -1,0 +1,18 @@
+@import "../../../../node_modules/matrix-react-sdk/res/themes/AlohaBlue/css/main.scss";
+// @import "_variables.scss"; // Overrides theme's variables at riot-web's stack
+@import "../../../../node_modules/matrix-react-sdk/res/css/_components.scss";
+@import "_common.scss";
+@import "_AppsDrawer.scss";
+@import "_BaseAvatar.scss";
+@import "_LeftPanel.scss";
+@import "_MatrixToolbar.scss";
+@import "_RoomSubList.scss";
+@import "_RoomTile.scss";
+@import "_SearchBox.scss";
+@import "_FilePanel.scss";
+@import "_EventTile.scss";
+@import "AlohaLogin.scss";
+
+
+
+
