--- res/themes/weheal/css/main.scss
@@ -1,4 +1,0 @@
-@import "../../../../node_modules/matrix-react-sdk/res/themes/AHNblue/css/main.scss";
-body {
-    background-color: #000;
-}
