+++ webpack.config.js
@@ -27,7 +27,7 @@ module.exports = {
         // CSS themes
         "theme-light":  "./node_modules/matrix-react-sdk/res/themes/light/css/light.scss",
         "theme-dark":   "./node_modules/matrix-react-sdk/res/themes/dark/css/dark.scss",
-        "theme-weheal":   "./res/themes/weheal/css/main.scss",
+        "theme-AlohaBlue":  "./res/themes/AlohaBlue/css/main.scss",
         "theme-status": "./res/themes/status/css/status.scss",
     },
     module: {
@@ -106,6 +106,10 @@ module.exports = {
             // same goes for js-sdk
             "matrix-js-sdk": path.resolve('./node_modules/matrix-js-sdk'),
         },
+        modules: [
+            'themes/weheal/',
+            'node_modules'
+        ]
     },
     externals: {
         // Don't try to bundle electron: leave it as a commonjs dependency
